==> Python_code/P01_survey.ipynb <==
# P01_survey 

### Program to obtain variable k, "Industry 4.0 Knowledge"

### This new variable encompasses knowledge of Information and Communication Technologies required by Industry 4.0.

1. Select questions 15 to 22 (see Appendix I), which correspond to the knowledge acquired on information technologies such as artificial Intelligence, big data, cybersecurity, software development and programming, networks and communications, robotics and automation, IoT, and cloud computing. These topics align with the study programs offered at the research center.

2. Based on an excerpt from Table 1 of the document, we selected a subset of the results corresponding to the mentioned technologies. This excerpt (Table 5 of the document) is contained in file "Tec_x_sector.csv".

3. The adoption percentages of the mentioned information technologies by the industrial sectors of the 26 countries were taken and normalized to be used as coefficients for the weighted sum.

4. Applying Equation 7 and 8, Table 6 was obtained. The normalized coefficients of all sectors for each information technology were obtained by applying Equation 9, and the results are shown in Table 7.


$$ sum_j = \sum_{i=0}^{n} p_i $$
$$ n_{ij} = \frac {p_i}{sum_j} , \, \{ i = 1, nit\} \, | \, j = 1, nsec \}$$

$$ C_i = \frac {\sum_j^{nsec}  n_{ij}}{nsec}, \, \{ i = 1, nit \} $$

   
5. The normalized values $Ci$ , represent the coefficients to obtain the new variable "IT Industry 4.0 Knowledge" (k), based on the weighted sum of the data received from the responses to questions 15 to 22 of the survey questionnaire, according to Equation 10.

$$ k = \sum_{i=1}^{nti} C_i  P_j \quad , \, \{ j = 15, 22 \}$$



### Reading the file containing an excerpt of Technologies will likely be adopted by 2025

import pandas as pd
import numpy as np
from pathlib import Path

# Reading the file containing an excerpt of Technologies will likely be adopted by 2025
# by the share of companies surveyed selected sectors (Table 5 of the document)
# DataFrame with 14 industrial' sectors (columns) and 7 information technologies (rows)

ruta_entrada = Path('..')/'Datasets'/'Tec_x_sector.csv'
data = pd.read_csv(ruta_entrada,delimiter=",",encoding = "ISO-8859-1")

### Obtaining the normalized coefficients

tecnologias = ['Artificial Intelligence', 'Big Data', 'Cybersecurity', 'IoT',
               'Robots', 'Cloud Computing', 'Software and Programming', 'Networks']


df = pd.DataFrame(data)

# Function to normalize values by columns (sectors)
def normalizar_por_columnas_y_calcular_Ci(df):
    # Normalize each column by dividing by the sum of the column
    df_normalizado = df.div(df.sum(axis=0), axis=1)
    
    # Add a new column with the normalized coefficients of all sectors (Ci) 
    df_normalizado['Ci'] = df_normalizado.mean(axis=1)
    return df_normalizado

# Normalize the DataFrame
df_normalizado = normalizar_por_columnas_y_calcular_Ci(df)

#Create a new DataFrame that represents Table 7. "Adoption percentages of IT across all industrial sectors and their normalized values
nuevo_df = pd.DataFrame({
    '(%)':df.mean(axis=1).round(0).astype(int),
    'Ci':df_normalizado['Ci']
})
  
nuevo_df.insert(0,"Technology adopted by all sectors", tecnologias)

# Display original and normalized DataFrames

# Configure pandas to display 9 decimal places in the visualization.
pd.set_option('display.float_format', '{:.9f}'.format)

print("Original DataFrame (adoption percentajes):")
print(df)
print("\n" + "="*50 + "\n")
print("Normalized DataFrame (normalized values by column):")
print(df_normalizado)
print("\n" + "="*50 + "\n")
print("DataFrame Table 7. Adoption percentages of IT across all industrial sectors and their normalized values")
print()
print(nuevo_df)


### Obtaining the new variable "IT Industry 4.0 Knowledge" (k)
&nbsp;
Based on the weighted sum of the data received from the responses to questions 15 to 22 of the survey questionnaire, according to Equation:
$$ k = \sum_{i=1}^{nti} C_i  P_j \quad , \, \{ j = 15, 22 \}$$


# Reading the file with the students' responses to the questionnaire.
ruta_entrada = Path('..')/'Datasets'/'enc_datosnum.csv'
data_enc = pd.read_csv(ruta_entrada,delimiter=",",encoding = "ISO-8859-1")
df_enc = pd.DataFrame(data_enc)

print("\n" + "="*50 + "\n")
print("DataFrame with the students' responses to the questionnaire")
print(df_enc)

# Select questions 15 to 22 from the responses DataFrame
preguntas_seleccionadas = df_enc.iloc[:, 15:23]  # Columnas 15 a 22 (índice base 0) 

KI4 = pd.DataFrame()
list_sumprod =[]
list_sumprod_rounded = []

# Obtaining k, weighted sum of the data received from the responses to questions 15 to 22 of the survey questionnaire
for i in range(len(df_enc)):
    # Multiplicar la columna 3 de Ci por pj de df_enc
    suma = sum(nuevo_df.iloc[j, 2] * preguntas_seleccionadas.iloc[i, j] for j in range(8))
    list_sumprod.append(suma)
    list_sumprod_rounded.append(round(suma - 0.2))

# Create a new DataFrame to store the results of the weighted sum k

KI4=pd.DataFrame({'k':list_sumprod,
                   'kr':list_sumprod_rounded})
print('------KI4------')
print(KI4)

#Export DataFrame KI4 as a file named KI4.csv.
ruta_salida = Path('..')/'Datasets'/'KI4.csv'
KI4.to_csv(ruta_salida, index=False)
