==> Python_code/P02_survey.ipynb <==
# P02_survey 

## Program to obtain descriptive statistics of the responses from the applied questionnaire.

### Create a new DataFrame from file with the students' responses to the questionnaire and file with variable "IT Industry 4.0 Knowledge" (k)


import numpy as np
import pandas as pd
from scipy import stats
from scipy import signal
from scipy.stats import norm
from scipy.stats import pearsonr
import statistics
import math
from pathlib import Path

ruta_entrada = Path('..')/'Datasets'/'enc_datosnum.csv'
df = pd.read_csv(ruta_entrada,delimiter=",",encoding = "ISO-8859-1")
data = pd.DataFrame(df)

ruta_entrada = Path('..')/'Datasets'/'KI4.csv'
df2 = pd.read_csv(ruta_entrada,delimiter=",",encoding = "ISO-8859-1")
dfki4 = pd.DataFrame(df2)

dconcat = pd.concat([data,dfki4], axis=1)

#Export DataFrame dconcat as a file named dconcat.csv.
ruta_salida = Path('..')/'Datasets'/'dconcat.csv'
dconcat.to_csv(ruta_salida, index=False)

dconcat

## Descriptive statistics

import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis

df = dconcat

# Function to obtain descriptive statistics
def calcular_estadisticas(df):
    resultados = {}
    
    for columna in df.columns:
        datos = df[columna]
        if columna == 'ï»¿K':
            continue
       
        estadisticas = {
            'Mean': datos.mean(),
            'Median': datos.median(),
            'Mode': datos.mode()[0],  # There can be multiple modes; we take the first one
            'Standard Deviation': datos.std(),
            'Variance': datos.var(),
            'Skewness': skew(datos),
            'Standar error of Skewness': np.sqrt(6 / len(datos)),  # Standard error approximation of skewness
            'Kurtosis': kurtosis(datos),
            'Range': datos.max() - datos.min()
        }
        
        resultados[columna] = estadisticas
    
    return pd.DataFrame(resultados)

# Obtain the descriptive statistics
estadisticas_df = calcular_estadisticas(df)

# Dispay descriptive statistics
print(estadisticas_df)
