==> Python_code/P03_survey.ipynb <==
# P03_survey 

### Program to obtain Frequencies and Percentajes of the responses from the applied questionnaire.

import pandas as pd
import numpy as np
from pathlib import Path

ruta_entrada = Path('..')/'Datasets'/'enc_datosnum.csv'
df = pd.read_csv(ruta_entrada,delimiter=",",encoding = "ISO-8859-1")
data = pd.DataFrame(df)


# Function to calculate the frequency and percentages of unique values in each column
def frecuencia_y_porcentajes(df):
    resultados = {}
    
    for columna in df.columns:
        # Calculate the frequency of unique values using value_counts()
        frecuencias = df[columna].value_counts().sort_index()
        
        # Calculate percentajes
        porcentajes = (frecuencias / len(df)) * 100
        
        # Almacenar los resultados en un DataFrame
        resultados[columna] = pd.DataFrame({
            'Frequency': frecuencias,
            'Percentaje (%)': porcentajes
        })
    
    return resultados

# Calculate the frequency and percentages
resultados = frecuencia_y_porcentajes(df)

# Display the results
for columna, resultado in resultados.items():
    print(f"Frequency and percentages for the column '{columna}':")
    print(resultado)
    print("\n" + "="*50 + "\n")
