==> Python_code/P04_survey.ipynb <==
# P04_survey 

### Program to perform the statistical normality test used to determine if a dataset follows a normal probability distribution 

import pandas as pd
import numpy as np
from scipy.stats import norm
from pathlib import Path

ruta_entrada = Path('..')/'Datasets'/'dconcat.csv'
data = pd.read_csv(ruta_entrada,delimiter=",",encoding = "ISO-8859-1")
df = pd.DataFrame(data)

# Create a new DataFrame to store the results
resultado = pd.DataFrame()

# Function to process each column
def procesar_columna(columna):
    
    # Columna 1: Values from 1 to 112
    resultado['col1'] = np.arange(1, 113)
    
    # Columna 2: Values in the column sorted from smallest to largest.


    resultado['col2'] = np.sort(columna)
    
    # Columna 3: Values in col1 * 2 - 1
    resultado['col3'] = resultado['col1'] * 2 - 1
    
    # Columna 4: Values in (col2 - mean de col2) / Standar deviation de col2
    media = resultado['col2'].mean()
    desviacion_estandar = resultado['col2'].std(ddof=1)
    print( media, desviacion_estandar)
    resultado['col4'] = (resultado['col2'] - media) / desviacion_estandar
    
    # Columna 5: Values in col4 reversed
    resultado['col5'] = resultado['col4'][::-1].reset_index(drop=True)
    
    # Columna 6: Cumulative distribution function (CDF) for each value in col2.
    resultado['col6'] = norm.cdf(resultado['col2'], loc=resultado['col2'].mean(), scale=resultado['col2'].std(ddof=1))
    
    # Columna 7: Values in col6 reversed
    resultado['col7'] = resultado['col6'][::-1].reset_index(drop=True)
    
    # Columna 8: Values of the natural logarithm (ln) of the values in col6.
    resultado['col8'] = np.log(resultado['col6'])
    
    # Columna 9: Values of the natural logarithm(1 - valores de col7)
    resultado['col9'] = np.log(1 - resultado['col7'])
    
    # Columna 10: Values of (1/112) * col3 * (col8 + col9)
    resultado['col10'] = (1 / 112) * resultado['col3'] * (resultado['col8'] + resultado['col9'])
    
    # Calculate the sum of the values in col10 (Si).

    Si = resultado['col10'].sum()
    
    adc = 112 + Si
    
    return resultado, Si, adc

# Process each column of the original DataFrame.

resultados_finales = {}
for columna in df.columns:
    resultados_finales[columna], Si, adc = procesar_columna(df[columna])
    print(f"For {columna}: Si = {Si:.4f}, adc = {adc:.4f}")
