==> Python_code/P12_survey.ipynb <==
# P12_survey

## Program to apply Left-tailed Student's t-test for the mean of variable kr (IT Industry 4.0 Knowledge)

One- tailed t-test for a single sample mean to the variable "IT Knowledge Industry 4.0"
* Variable (k) Industry 4.0 Knowledge
* The confidence interval is 95%, with a significance level of $ \alpha= 0.05 $
* Ho: media >= 4 Students are expected to consider the knowledge acquired during their studies on the IT adopted by Industry 4.0 as sufficient or totally sufficient.
* Ha: media < 4
* The test statistics: $ t = \sqrt{n} \frac{\bar{x} - \mu}{s} $

import numpy as np
import pandas as pd
from scipy import stats
from scipy.stats import norm
import statistics
import math
from pathlib import Path

# Data
ruta_entrada = Path('..')/'Datasets'/'KI4.csv'
data = pd.read_csv(ruta_entrada,delimiter=",",encoding = "ISO-8859-1")
df = pd.DataFrame(data)

# 95% confidence level
nivel_confianza = 0.95

# Get the value of z (critical value)
# For a 95% confidence level, the area in the left tail is (1 - confidence_level) / 2
z = norm.ppf(1 - (1 - nivel_confianza) / 2)
print(f"The value of z for a 95% confidence level is: {z:.6f}")

# Get the mean and standard deviation from the sample.
datos = df['kr']
media = datos.mean()
desviacion_estandar = datos.std()
print(f"The value of the mean is: {media:2.6f} \nThe value of the standard deviation is: {desviacion_estandar:2.6f}")

# Get the value of t
mu = 4 # value of the mean to test
n = len(datos)
t = math.sqrt(n) * (media - mu)/desviacion_estandar
print(f"The value of t is: {t:2.6f}\n")

# t test
if t < z:
    print(f"Given that: {t:2.6f} < {z:2.6f}\n\n"
             + "There is sufficient evidence to reject the null hypothesis Ho, meaning that students perceive that:\n"
             + "the level of knowledge acquired during their master's studies regarding the IT skills needed to\n"
             + "integrate into Industry 4.0 is insufficient.\n"
             + "Therefore, the alternative hypothesis Ha is accepted.")
else:
    print(f"Given that: {t:2.6f} >= {z:2.6f}\n\n"
             + "There is sufficient evidence to accept the null hypothesis Ho, meaning that students perceive that:\n"
             + "the level of knowledge acquired during their master's studies regarding the IT skills needed to\n"
             + "integrate into Industry 4.0 is sufficient.")
